feat: restructure project & update includes
