+++ src/assets/folder_empty.png
