+++ src/gui/builders/modal_builder.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <QInputDialog>
 #include <QDialog>
 #include <QInputDialog>
