+++ src/main.cpp
@@ -1,5 +1,6 @@
 #include <QApplication>
-#include "file_explorer.h"
+
+#include "gui/file_explorer.h"
 
 int main(int argc, char *argv[]) {
   QApplication app(argc, argv);
